src/requirements-dev.txt
django-debug-toolbar
django-extensions
django-pdb
ipython
